+++ tests/org/intellij/grammar/BnfGeneratorTest.java
@@ -4,8 +4,6 @@
 
 package org.intellij.grammar;
 
-import com.intellij.idea.Bombed;
-
 public class BnfGeneratorTest extends BnfGeneratorAbstractTest {
   public BnfGeneratorTest() {
     super("generator");
@@ -32,8 +30,8 @@ public class BnfGeneratorTest extends BnfGeneratorAbstractTest {
   public void testConsumeMethods() throws Exception { doGenTest(false); }
   public void testGenOptions() throws Exception { doGenTest(true); }
 
-  @Bombed(year = 2030, user = "author", month = 1, day = 1, description = "not implemented")
-  public void testUpperRules() throws Exception { doGenTest(true); }
+  // TODO not implemented
+  public void _testUpperRules() throws Exception { doGenTest(true); }
   public void testFixes() throws Exception { doGenTest(true); }
 
   public void testEmpty() throws Exception {