+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 * JFlex: update library from jflex-1.9.1 to jflex-1.9.2 #349
+* Update IntelliJ Platform to `2023.3`
 * Compatibility: IntelliJ IDEA 2024.2
 * Context menu: Fleet: Generate Parser Code
 * Context menu: Fleet: Generate JFlex Lexer